requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
autopep8==2.0.4
bcrypt==4.1.2
blinker==1.7.0
boto3==1.33.10
botocore==1.33.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
crispy-bootstrap5==2023.10
defusedxml==0.7.1
distlib==0.3.7
Django==5.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==24.2
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.13.1
flake8==7.0.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
httpie==3.2.2
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.0
npm==0.1.1
openpyxl==3.1.2
optional-django==0.1.0
packaging==24.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
pwinput==1.0.3
pycodestyle==2.11.1
pycryptodome==3.21.0
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pypiwin32==223
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
s3transfer==0.8.2
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.25.0
Werkzeug==3.0.1
